requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
yfinance==0.2.28
python-dateutil==2.8.2
psycopg2-binary==2.9.9
python-multipart==0.0.6